pyproject.toml
[project]
name = "graph-rag-cookbook"
version = "0.1.0"
description = "Self-hosted, reusable framework for constructing an LLM-queryable knowledge graph over a markdown corpus, publishing portable graph artifacts for downstream consumer agents."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "graph-RAG_CookBook contributors" }]
keywords = ["knowledge-graph", "rag", "llm", "obsidian", "falkordb", "qdrant"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

# Runtime deps are intentionally minimal in Sprint 1 — packages are scaffolding only.
# Sprint 2+ will pin extraction/graph/vector/emitter dependencies.
dependencies = [
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "pyyaml>=6.0",
    "structlog>=24.1",
    "typer>=0.12",
    "jsonschema>=4.22",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-pyyaml>=6.0",
    "types-jsonschema>=4.22",
]

[project.scripts]
kg = "knowledge_graph.cli:app"

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/knowledge_graph"]

# ─────────────────────────────────────────────────────────────────────
# Ruff — lint + format
# ─────────────────────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "tests", "evals"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "W",    # pycodestyle warnings
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific
]
ignore = [
    "E501",  # line-too-long (formatter handles)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ─────────────────────────────────────────────────────────────────────
# Mypy — type checking
# ─────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
no_implicit_reexport = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
files = ["src", "tests", "evals"]

[[tool.mypy.overrides]]
module = ["tests.*", "evals.harness.tests.*"]
disallow_untyped_defs = false

# ─────────────────────────────────────────────────────────────────────
# Pytest
# ─────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests", "evals/harness"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--showlocals",
]
asyncio_mode = "auto"
filterwarnings = ["error"]

[tool.coverage.run]
source = ["src/knowledge_graph", "evals/harness"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
